Changed scope + Item Class
Changed the scope and direction of the game (entirely). Instead of a thief, you'll be in a life is strange sort of setting, where you play as a psychic getting ready for the day.

All that aside, added an item class that contains an empty string and an empty function