Improved movement, added some notes
Improved and added to the movement section of the script. Also added a quick thing to allow for you to check spells in the same constructor.

Also added a function to check if a room has a wall between another room.

I also added some notes to add some clarity to things as well as the new function.